*Added existing components usage instead of mocked one to collection settings page (card display mode);
* Renamed chips list and images list components to have plural form;
* Renamed cards status interface to make it more correct naming;
* Added class for disabling ripple effect in ion item;
