+++ src/app/features/collection/components/cards-list/cards-list.component.html
