+++ src/app/features/collection/components/chips-list/chips-list.component.html
@@ -1,7 +1,8 @@
 <div class="tw-flex tw-flex-wrap tw-justify-center tw-py-1">
   @for (chip of chipList(); track chip.id) {
     <app-chip
-      [chip]="chip"
+      class="tw-m-1"
+      [card]="chip"
       [isLoading]="getLoadingStatus(chipsPendingResponse(), chip.id)"
       (clickChip)="handleChipClick($event, chip)" />
   }
