+++ src/app/models/collection.models.ts
@@ -1,7 +1,7 @@
 export interface Card {
   id: string;
   rarity: string;
-  status: StatusCard;
+  status: CardStatus;
   character_name: string;
   serial_number: string;
   image_url: string;
@@ -9,10 +9,10 @@ export interface Card {
 
 export interface UpdateCardDto {
   id: string;
-  status: StatusCard;
+  status: CardStatus;
 }
 
-export enum StatusCard {
+export enum CardStatus {
   Collected = 'collected',
   NotCollected = 'not_collected',
   NotExisting = 'not_existing',
