+++ src/app/pages/private/collection/collection.page.html
@@ -20,9 +20,9 @@
 
   <ion-content>
     @if(cardsDisplayMode() === cardsDisplayModeEnum.Chip) {
-    <app-chip-list />
+    <app-chips-list />
     } @else {
-    <app-card-list />
+    <app-cards-list />
     }
   </ion-content>
